+++ aesop/commands/aspects/__init__.py
@@ -1,0 +1,5 @@
+from .user_defined_resources import tags_app
+
+__all__ = [
+    "tags_app",
+]
