+++ aesop/commands/aspects/user_defined_resources/__init__.py
@@ -1,0 +1,3 @@
+from .tags.tags import app as tags_app
+
+__all__ = ["tags_app"]
