+++ aesop/commands/aspects/user_defined_resources/tags/commands/add.py
@@ -1,8 +1,7 @@
-from typing import List, Optional
+from typing import List
 
-from aesop.commands.tags.models import BatchAddTagsInput, GovernedTag
+from aesop.commands.aspects.user_defined_resources.tags.models import GovernedTag
 from aesop.config import AesopConfig
-from aesop.console import console
 from aesop.graphql.generated.enums import UserDefinedResourceType
 from aesop.graphql.generated.input_types import (
     UserDefinedResourceDescriptionInput,
@@ -11,7 +10,7 @@ from aesop.graphql.generated.input_types import (
 )
 
 
-def _add_tags(
+def add_tags(
     tags: List[GovernedTag],
     config: AesopConfig,
 ) -> List[str]:
@@ -35,23 +34,3 @@ def _add_tags(
         raise ValueError
     created_ids = [res.id for res in resp.create_user_defined_resource]
     return created_ids
-
-
-def add(
-    name: str,
-    description: Optional[str],
-    config: AesopConfig,
-) -> None:
-    tag = GovernedTag(name=name, description=description)
-    created_ids = _add_tags([tag], config)
-    console.ok("Successfully created governed tag")
-    console.print(created_ids[0])
-
-
-def batch_add(
-    batch_add_tags_input: BatchAddTagsInput,
-    config: AesopConfig,
-) -> None:
-    created_ids = _add_tags(batch_add_tags_input.tags, config)
-    console.ok("Successfully created governed tags")
-    console.print(created_ids)
