+++ aesop/commands/aspects/user_defined_resources/tags/commands/remove.py
@@ -1,14 +1,11 @@
 from typing import List
 
-from rich import print
-
-from aesop.commands.tags.models import BatchRemoveTagsInput
 from aesop.config import AesopConfig
 from aesop.graphql.generated.input_types import UserDefinedResourceDeleteInput
 from aesop.graphql.generated.remove_governed_tags import RemoveGovernedTags
 
 
-def _remove_tags(
+def remove_tags(
     tag_ids: List[str],
     config: AesopConfig,
 ) -> RemoveGovernedTags:
@@ -18,19 +15,3 @@ def _remove_tags(
             ids=tag_ids,
         )
     )
-
-
-def remove(
-    tag_id: str,
-    config: AesopConfig,
-) -> None:
-    _remove_tags([tag_id], config)
-    print(f"Removed tag {tag_id}")
-
-
-def batch_remove(
-    input: BatchRemoveTagsInput,
-    config: AesopConfig,
-) -> None:
-    resp = _remove_tags(input.tag_ids, config)
-    print(resp.model_dump_json(indent=2))
